Issue #56: allow passing a parameterless function to call().
For this, we introduce a `ThrowingRunnable` interface analogously to `ThrowingFunction` and `ThrowingConsumer`,
and simply call it in the Executor, ignoring the response and/or any entities.

Tests were adapted to make everything 100% – I'm not sure that this coverage number is actually meaningful in this case. (It looks like some of the previous `call`s were only tested for either synchronous or asynchronous operation, not for both.)