Avoid threads being scheduled twice
Currently threads are being scheduled twice on their request executor.

* at the very start of a Riptide request (required for isolation)
* after the network phase (intented to force execution back on request
executor)

The second handover was introduced with NIO to accomodate for request
factories which use their own pool. The benefit of forcing execution
back on the request executor is questionable, as any IO is done by
definition after the network phase.

Removing the handover prevents the same excution being scheduled twice
onto the same request executor. In high pressure situations (more
requests than configured threads on a single client), this leads to
deadlock situations in which running executions cannot leave the network
phase, as all other threads are occupied with new executions waiting for
a free slot to start.